requirements.txt
gdal
keras
tensorflow
matplotlib
Boto3
tqdm
Pillow
yaml
skfmm
osgeo